Relocated to src/test-utils.
